tests/qdrant_llamaindex/requirements.txt
llama-index-core
llama-index-vector-stores-qdrant
llama-index-embeddings-ollama
qdrant-client
